+++ R/dark_mode.R
@@ -1,8 +1,10 @@
-#' Dark mode with right legend
+#' Dark mode theme with right legend
 #'
-#' @description Dark mode with right legend using `darkness`and `linewidthness`.
+#' @description Dark mode theme with right legend using `darkness` and `linewidthness`.
 #'
-#' @inheritParams create_mode_r
+#' @param text_size The base size of the text. Defaults to 11. The title is 110% of this, caption 85% and tag 120%.
+#' @param text_family The family of the text. Defaults to "".
+#' @param ... Provided to support trailing commas only.
 #'
 #' @return A ggplot theme.
 #' @export
@@ -22,22 +24,30 @@
 #'   )
 #'
 dark_mode_r <- function (
-    size = 11,
-    family = "") {
+    text_size = 11,
+    text_family = "",
+    ...
+) {
 
-  create_mode_r(
-    size = size,
-    family = family,
-    col_palette = darkness,
-    linewidth_palette = linewidthness
+  get_mode_r(
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = darkness[1],
+    axis_line_colour = darkness[2],
+    axis_line_linewidth = linewidthness[1],
+    panel_line_colour = darkness[3],
+    panel_line_linewidth = linewidthness[2],
+    panel_background_colour = darkness[4],
+    plot_background_colour = darkness[5],
+    ...
   )
 }
 
-#' Dark mode with top legend
+#' Dark mode theme with top legend
 #'
-#' @description Dark mode with top legend using `darkness`and `linewidthness`.
+#' @description Dark mode theme with top legend using `darkness` and `linewidthness`.
 #'
-#' @inheritParams create_mode_t
+#' @inheritParams dark_mode_r
 #'
 #' @return A ggplot theme.
 #' @export
@@ -57,22 +67,30 @@ dark_mode_r <- function (
 #'   )
 #'
 dark_mode_t <- function (
-    size = 11,
-    family = "") {
+    text_size = 11,
+    text_family = "",
+    ...
+) {
 
-  create_mode_t(
-    size = size,
-    family = family,
-    col_palette = darkness,
-    linewidth_palette = linewidthness
+  get_mode_t(
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = darkness[1],
+    axis_line_colour = darkness[2],
+    axis_line_linewidth = linewidthness[1],
+    panel_line_colour = darkness[3],
+    panel_line_linewidth = linewidthness[2],
+    panel_background_colour = darkness[4],
+    plot_background_colour = darkness[5],
+    ...
   )
 }
 
-#' Dark mode with bottom legend
+#' Dark mode theme with bottom legend
 #'
-#' @description Dark mode with bottom legend using `darkness`and `linewidthness`.
+#' @description Dark mode theme with bottom legend using `darkness` and `linewidthness`.
 #'
-#' @inheritParams create_mode_b
+#' @inheritParams dark_mode_r
 #'
 #' @return A ggplot theme.
 #' @export
@@ -92,22 +110,30 @@ dark_mode_t <- function (
 #'   )
 #'
 dark_mode_b <- function (
-    size = 11,
-    family = "") {
+    text_size = 11,
+    text_family = "",
+    ...
+) {
 
-  create_mode_b(
-    size = size,
-    family = family,
-    col_palette = darkness,
-    linewidth_palette = linewidthness
+  get_mode_b(
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = darkness[1],
+    axis_line_colour = darkness[2],
+    axis_line_linewidth = linewidthness[1],
+    panel_line_colour = darkness[3],
+    panel_line_linewidth = linewidthness[2],
+    panel_background_colour = darkness[4],
+    plot_background_colour = darkness[5],
+    ...
   )
 }
 
-#' Dark mode with no legend
+#' Dark mode theme with no legend
 #'
-#' @description Dark mode with no legend using `darkness`and `linewidthness`.
+#' @description Dark mode theme with no legend using `darkness` and `linewidthness`.
 #'
-#' @inheritParams create_mode_n
+#' @inheritParams dark_mode_r
 #'
 #' @return A ggplot theme.
 #' @export
@@ -127,13 +153,21 @@ dark_mode_b <- function (
 #'   )
 #'
 dark_mode_n <- function (
-    size = 11,
-    family = "") {
+    text_size = 11,
+    text_family = "",
+    ...
+) {
 
-  create_mode_n(
-    size = size,
-    family = family,
-    col_palette = darkness,
-    linewidth_palette = linewidthness
+  get_mode_n(
+    text_size = text_size,
+    text_family = text_family,
+    text_colour = darkness[1],
+    axis_line_colour = darkness[2],
+    axis_line_linewidth = linewidthness[1],
+    panel_line_colour = darkness[3],
+    panel_line_linewidth = linewidthness[2],
+    panel_background_colour = darkness[4],
+    plot_background_colour = darkness[5],
+    ...
   )
 }